File Class implemented

+++ backend/src/main/java/com/smu/oopg1t4/util/SequenceGeneratorService.java
@@ -1,6 +1,6 @@
 package com.smu.oopg1t4.util;
 
-import com.example.demo.util.DatabaseSequence;
+import com.smu.oopg1t4.util.DatabaseSequence;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.mongodb.core.MongoOperations;
 import org.springframework.data.mongodb.core.query.Update;
@@ -24,7 +24,7 @@ public class SequenceGeneratorService {
         this.mongoOperations = mongoOperations;
     }
 
-    public long generateSequence(String seqName) {
+    public int generateSequence(String seqName) {
 
         DatabaseSequence counter = mongoOperations.findAndModify(query(where("_id").is(seqName)),
                 new Update().inc("seq",1), options().returnNew(true).upsert(true),
